+++ react-spa-msal-entra-id-quickstart/components/providers/ClientFluentProvider.tsx
